requirements.txt
google-genai
python-dotenv
pypdf
rank_bm25
